CORE-173: Swift interface abstracting Bitcoin+Ethereum+<others>

+++ ethereum/base/BREthereumBase.h
@@ -73,7 +73,7 @@ BRSetFreeAll (BRSet *set, BRSetItemFree itemFree) {
 #include "BREthereumSignature.h"
 
 typedef uint64_t BREthereumTimestamp;  // A Unix time
-#define ETHEREUM_TIMESTAMP_UNKNOWN    (0)
+#define ETHEREUM_TIMESTAMP_UNKNOWN    ((uint64_t) 0)
 
 typedef enum {
     RLP_TYPE_NETWORK,
